split out main.ts
